OFH-Dashboard/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SocketIO==5.3.6
python-socketio==5.11.0
pandas==2.2.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
kafka-python==2.2.15
six>=1.16.0
PyJWT==2.8.0
Werkzeug==3.0.1
Flask-Limiter==3.5.0
redis>=4.0.0
tenacity==8.2.3
requests==2.31.0
jsonschema==4.25.1
bcrypt==4.0.1
psutil==5.9.8
Flask-Compress==1.14
flasgger==0.9.7.1

